feat/fix (projectDetails): projectDetails component projectOption section  splitted to projectOption component. projectOptions added for others bottomSheet
