+++ src/frontend/src/components/ProjectDetails/ProjectOptions.tsx
@@ -1,0 +1,161 @@
+import Accordion from '../../components/common/Accordion';
+import React, { useState } from 'react';
+import CoreModules from '../../shared/CoreModules';
+import AssetModules from '../../shared/AssetModules';
+import environment from '../../environment';
+import { DownloadDataExtract, DownloadProjectForm } from '../../api/Project';
+import MapLegends from '../../components/MapLegends';
+
+const ProjectOptions = ({ setToggleGenerateModal }) => {
+  const dispatch = CoreModules.useAppDispatch();
+  const params = CoreModules.useParams();
+
+  const [toggleAction, setToggleAction] = useState(false);
+  const downloadProjectFormLoading = CoreModules.useAppSelector((state) => state.project.downloadProjectFormLoading);
+  const downloadDataExtractLoading = CoreModules.useAppSelector((state) => state.project.downloadDataExtractLoading);
+  const defaultTheme = CoreModules.useAppSelector((state) => state.theme.hotTheme);
+
+  const encodedId = params.id;
+  const decodedId = environment.decode(encodedId);
+
+  const handleDownload = (downloadType) => {
+    if (downloadType === 'form') {
+      dispatch(
+        DownloadProjectForm(`${import.meta.env.VITE_API_URL}/projects/download_form/${decodedId}/`, downloadType),
+      );
+    } else if (downloadType === 'geojson') {
+      dispatch(
+        DownloadProjectForm(`${import.meta.env.VITE_API_URL}/projects/${decodedId}/download_tasks`, downloadType),
+      );
+    }
+  };
+  const onDataExtractDownload = () => {
+    dispatch(
+      DownloadDataExtract(`${import.meta.env.VITE_API_URL}/projects/features/download/?project_id=${decodedId}`),
+    );
+  };
+  return (
+    <div className="fmtm-mt-4">
+      <div>
+        <div
+          className={`fmtm-flex fmtm-gap-5 fmtm-py-4 sm:fmtm-hidden fmtm-justify-between  fmtm-items-center fmtm-mx-7 fmtm-mb-2 ${
+            toggleAction ? 'fmtm-border-b-[#929DB3] fmtm-border-b-[1px]' : ''
+          }`}
+        >
+          <p className="fmtm-text-xl fmtm-italic">Project Options</p>
+          <div
+            className={
+              'fmtm-rounded-full fmtm-shadow-gray-400 fmtm-w-8 fmtm-h-8 fmtm-flex fmtm-justify-center fmtm-items-center fmtm-shadow-lg fmtm-cursor-pointer'
+            }
+            onClick={() => setToggleAction(!toggleAction)}
+          >
+            <AssetModules.ArrowRightIcon
+              color=""
+              style={{ fontSize: 32 }}
+              className={`${toggleAction ? 'fmtm-rotate-90' : ''}`}
+            />
+          </div>
+        </div>
+      </div>
+      <div
+        className={`fmtm-flex fmtm-flex-col lg:fmtm-flex-row fmtm-gap-6 lg:fmtm-gap-0 fmtm-px-3 sm:fmtm-px-0 ${
+          toggleAction ? '' : 'fmtm-hidden sm:fmtm-flex'
+        }`}
+      >
+        <div className="fmtm-w-full fmtm-flex fmtm-flex-col fmtm-items-start sm:fmtm-flex-row  sm:fmtm-justify-center lg:fmtm-justify-start sm:fmtm-items-center fmtm-gap-6 fmtm-ml-4">
+          <CoreModules.LoadingButton
+            onClick={() => handleDownload('form')}
+            sx={{ width: 'unset' }}
+            loading={downloadProjectFormLoading.type === 'form' && downloadProjectFormLoading.loading}
+            loadingPosition="end"
+            endIcon={<AssetModules.FileDownloadIcon />}
+            variant="contained"
+            color="error"
+          >
+            Form
+          </CoreModules.LoadingButton>
+          <CoreModules.LoadingButton
+            onClick={() => handleDownload('geojson')}
+            sx={{ width: 'unset' }}
+            loading={downloadProjectFormLoading.type === 'geojson' && downloadProjectFormLoading.loading}
+            loadingPosition="end"
+            endIcon={<AssetModules.FileDownloadIcon />}
+            variant="contained"
+            color="error"
+          >
+            Tasks
+          </CoreModules.LoadingButton>
+          <CoreModules.LoadingButton
+            onClick={() => onDataExtractDownload()}
+            sx={{ width: 'unset' }}
+            loading={downloadDataExtractLoading}
+            loadingPosition="end"
+            endIcon={<AssetModules.FileDownloadIcon />}
+            variant="contained"
+            color="error"
+            className="fmtm-truncate"
+          >
+            Data Extract
+          </CoreModules.LoadingButton>
+        </div>
+        <div className="fmtm-flex fmtm-flex-col sm:fmtm-flex-row sm:fmtm-justify-center lg:fmtm-justify-end fmtm-w-full fmtm-ml-4 fmtm-gap-6">
+          <CoreModules.Link
+            to={`/projectInfo/${encodedId}`}
+            style={{
+              display: 'flex',
+              justifyContent: 'flex-end',
+              alignItems: 'flex-end',
+              textDecoration: 'none',
+              marginRight: '15px',
+            }}
+            className="fmtm-w-fit"
+          >
+            <CoreModules.Button variant="contained" color="error">
+              ProjectInfo
+            </CoreModules.Button>
+          </CoreModules.Link>
+          <CoreModules.Button
+            onClick={() => setToggleGenerateModal(true)}
+            variant="contained"
+            color="error"
+            sx={{ width: '200px', mr: '15px' }}
+            endIcon={<AssetModules.BoltIcon />}
+            className="fmtm-truncate"
+          >
+            Generate MbTiles
+          </CoreModules.Button>
+          <CoreModules.Link
+            to={`/edit-project/project-details/${encodedId}`}
+            style={{
+              display: 'flex',
+              justifyContent: 'flex-end',
+              alignItems: 'flex-end',
+              textDecoration: 'none',
+              marginRight: '15px',
+            }}
+            className="fmtm-w-fit"
+          >
+            <CoreModules.Button variant="outlined" color="error" className="fmtm-truncate">
+              Edit Project
+            </CoreModules.Button>
+          </CoreModules.Link>
+        </div>
+        <div className="fmtm-px-1 sm:fmtm-hidden">
+          <Accordion
+            collapsed={true}
+            disableHeaderClickToggle
+            onToggle={() => {}}
+            header={<div className="fmtm-text-[#2C3038] fmtm-font-bold fmtm-text-xl">Map Legends</div>}
+            body={
+              <div className="fmtm-mt-4">
+                <MapLegends defaultTheme={defaultTheme} />
+              </div>
+            }
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ProjectOptions;
